feat: resolve E2E booking flow, fix Kafka deserialization, and improve test maintainability
- Fix Kafka deserialization by configuring JsonDeserializer type mappings in booking, charting, and billing services.

- Resolve Spring Boot 3 parameter reflection issues by compiling with -parameters and naming path/request parameters.

- Fix Flyway migration type mismatch for day_of_week and seed default clinic/practitioner data.

- Correct booking-service port mapping in docker-compose.yml.

- Stabilize frontend date query parameters using useMemo in BookingForm to prevent infinite loading spinner.

- Resolve OptimisticLockException on TimeSlot by removing @Version annotation to avoid conflict with pessimistic write lock.

- Introduce data-testid attributes to booking components and refactor E2E test to use robust semantic selectors.

+++ implimantation/booking-service/src/main/resources/db/migration/V1__init.sql
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS availability_templates (
     template_id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     tenant_id UUID NOT NULL,
     practitioner_id UUID NOT NULL REFERENCES practitioners(practitioner_id),
-    day_of_week SMALLINT NOT NULL CHECK (day_of_week BETWEEN 0 AND 6),
+    day_of_week INTEGER NOT NULL CHECK (day_of_week BETWEEN 0 AND 6),
     start_time TIME NOT NULL,
     end_time TIME NOT NULL,
     slot_duration_minutes INTEGER NOT NULL DEFAULT 30,
